fix：新的加载界面显示过长的问题
