+++ html/js/newtab.js
@@ -4,11 +4,36 @@
  */
 
 import { initI18n, getI18nMessage } from './modules/i18n.js';
-import { initBackgroundImage, setupBackgroundEvents } from './modules/backgroundImage.js';
-import { initSearchEngine, setupSearchEvents } from './modules/searchEngine.js';
-import { initBookmarks, setupBookmarkEvents } from './modules/bookmarks.js';
-import { initIconManager, preloadIcons } from './modules/iconManager.js';
-import { initClock } from './modules/clockWidget.js';
+import { 
+    initBackgroundImage, 
+    setupBackgroundEvents,
+    refreshBackgroundImage
+} from './modules/backgroundImage.js';
+import { 
+    initSearchEngine, 
+    setupSearchEvents 
+} from './modules/searchEngine.js';
+import { 
+    initBookmarks, 
+    setupBookmarkEvents 
+} from './modules/bookmarks.js';
+import { 
+    initIconManager, 
+    preloadIcons 
+} from './modules/iconManager.js';
+import { 
+    initClock, 
+    updateClock 
+} from './modules/clockWidget.js';
+import { 
+    showLoadingIndicator,
+    hideLoadingIndicator,
+    updateLoadingProgress,
+    showErrorMessage,
+    showNotification,
+    initUIEvents,
+    handleKeyDown
+} from './modules/utils.js';
 
 // 当前版本号
 const VERSION = '1.1.5';
@@ -72,8 +97,6 @@ async function init() {
             console.error('Failed to initialize icon manager:', error);
             throw new Error('图标资源加载失败: ' + error.message);
         }
-        
-        // 继续为每个模块添加类似的错误处理...
 
         try {
             // 3. 初始化背景图像
@@ -148,12 +171,14 @@ async function init() {
         isInitialized = true;
         console.log('New tab page initialization complete');
         
+        // 立即隐藏加载界面
+        hideLoadingIndicator();
+        
     } catch (error) {
         // 处理初始化错误
         console.error('Failed to initialize new tab page:', error);
         showErrorMessage('初始化失败，请刷新页面重试。');
-    } finally {
-        // 无论成功或失败，都隐藏加载指示器
+        // 确保在错误情况下也隐藏加载界面
         hideLoadingIndicator();
     }
 }
@@ -172,137 +197,7 @@ function setupEvents() {
     setupBookmarkEvents();
     
     // 设置通用页面事件
-    setupGeneralPageEvents();
-    
-    // 初始化模态框通用事件
-    initModalEvents();
-}
-
-/**
- * 设置通用页面事件
- */
-function setupGeneralPageEvents() {
-    // 处理页面加载完成事件
-    window.addEventListener('load', handlePageLoad);
-    
-    // 处理页面调整大小事件
-    window.addEventListener('resize', handleWindowResize);
-    
-    // 全局键盘快捷键
-    document.addEventListener('keydown', handleKeyDown);
-    
-    // 处理点击事件，关闭弹出菜单等
-    document.addEventListener('click', handleDocumentClick);
-}
-
-/**
- * 初始化模态框事件
- */
-function initModalEvents() {
-    // 获取所有模态框
-    const modals = document.querySelectorAll('.modal');
-    
-    modals.forEach(modal => {
-        // 关闭按钮
-        const closeButtons = modal.querySelectorAll('.modal-close');
-        closeButtons.forEach(button => {
-            button.addEventListener('click', () => {
-                modal.style.display = 'none';
-            });
-        });
-        
-        // 点击模态框外部关闭
-        modal.addEventListener('click', (e) => {
-            if (e.target === modal) {
-                modal.style.display = 'none';
-            }
-        });
-    });
-}
-
-/**
- * 处理文档点击事件
- * @param {Event} e - 事件对象
- */
-function handleDocumentClick(e) {
-    // 关闭下拉菜单
-    const dropdowns = document.querySelectorAll('.dropdown-menu.active');
-    dropdowns.forEach(dropdown => {
-        if (!dropdown.parentElement.contains(e.target)) {
-            dropdown.classList.remove('active');
-        }
-    });
-}
-
-/**
- * 处理页面加载完成事件
- */
-function handlePageLoad() {
-    // 移除加载屏幕
-    const loadingScreen = document.getElementById('loading-screen');
-    if (loadingScreen) {
-        loadingScreen.style.opacity = '0';
-        setTimeout(() => {
-            loadingScreen.style.display = 'none';
-        }, 500); // 500ms过渡动画后隐藏
-    }
-    
-    // 聚焦搜索框
-    const searchInput = document.getElementById('search-input');
-    if (searchInput) {
-        setTimeout(() => {
-            searchInput.focus();
-        }, 100);
-    }
-}
-
-/**
- * 处理窗口调整大小事件
- */
-function handleWindowResize() {
-    // 在这里处理响应式布局的调整
-    // 例如根据窗口大小调整元素位置、大小等
-}
-
-/**
- * 处理键盘按键事件
- * @param {KeyboardEvent} e - 键盘事件对象
- */
-function handleKeyDown(e) {
-    // 处理ESC键 - 关闭模态框/面板
-    if (e.key === 'Escape') {
-        // 关闭所有模态框
-        document.querySelectorAll('.modal').forEach(modal => {
-            modal.style.display = 'none';
-        });
-        
-        // 关闭设置面板
-        const settingsPanel = document.getElementById('settings-panel');
-        if (settingsPanel) {
-            settingsPanel.classList.remove('active');
-        }
-        
-        // 关闭搜索建议
-        const searchSuggestions = document.getElementById('search-suggestions');
-        if (searchSuggestions) {
-            searchSuggestions.innerHTML = '';
-        }
-    }
-    
-    // 处理/键 - 聚焦搜索框
-    if (e.key === '/' && !e.target.matches('input, textarea')) {
-        e.preventDefault();
-        document.getElementById('search-input')?.focus();
-    }
-    
-    // 处理Alt+S - 打开设置
-    if (e.key === 's' && e.altKey) {
-        e.preventDefault();
-        const settingsPanel = document.getElementById('settings-panel');
-        if (settingsPanel) {
-            settingsPanel.classList.toggle('active');
-        }
-    }
+    initUIEvents();
 }
 
 /**
@@ -379,105 +274,6 @@ async function cleanupCacheData() {
 }
 
 /**
- * 显示加载指示器
- */
-function showLoadingIndicator() {
-    const loadingScreen = document.getElementById('loading-screen');
-    if (loadingScreen) {
-        // 创建或获取进度条元素
-        let progressBar = document.getElementById('loading-progress-bar');
-        if (!progressBar) {
-            progressBar = document.createElement('div');
-            progressBar.id = 'loading-progress-bar';
-            progressBar.style.cssText = 'width:0%;height:4px;background:linear-gradient(90deg,#4285f4,#34a853,#fbbc05,#ea4335);position:absolute;bottom:0;left:0;transition:width 0.3s;border-radius:0 2px 2px 0;';
-            loadingScreen.appendChild(progressBar);
-        }
-        
-        // 创建或获取状态文本元素
-        let statusText = document.getElementById('loading-status-text');
-        if (!statusText) {
-            statusText = document.createElement('div');
-            statusText.id = 'loading-status-text';
-            statusText.style.cssText = 'margin-top:10px;font-size:14px;color:#666;';
-            statusText.textContent = '准备加载...';
-            loadingScreen.appendChild(statusText);
-        }
-        
-        // 重置进度
-        progressBar.style.width = '0%';
-        statusText.textContent = '准备加载...';
-        
-        loadingScreen.style.display = 'flex';
-    }
-}
-
-/**
- * 隐藏加载指示器
- */
-function hideLoadingIndicator() {
-    console.log('Hiding loading indicator');
-    const loadingScreen = document.getElementById('loading-screen');
-    if (loadingScreen) {
-        loadingScreen.style.opacity = '0';
-        setTimeout(() => {
-            loadingScreen.style.display = 'none';
-            console.log('Loading screen hidden');
-        }, 500);
-    }
-}
-
-/**
- * 更新加载进度
- * @param {number} percent - 加载百分比(0-100)
- * @param {string} statusMessage - 当前状态消息
- */
-function updateLoadingProgress(percent, statusMessage) {
-    const progressBar = document.getElementById('loading-progress-bar');
-    const statusText = document.getElementById('loading-status-text');
-    
-    if (progressBar) {
-        progressBar.style.width = `${percent}%`;
-    }
-    
-    if (statusText && statusMessage) {
-        statusText.textContent = statusMessage;
-    }
-}
-
-/**
- * 显示错误消息
- * @param {string} message - 错误消息
- */
-function showErrorMessage(message) {
-    const errorContainer = document.getElementById('error-container');
-    if (!errorContainer) {
-        // 如果不存在错误容器，创建一个
-        const container = document.createElement('div');
-        container.id = 'error-container';
-        container.style.cssText = 'position:fixed;top:10px;right:10px;background:#f44336;color:white;padding:10px 20px;border-radius:5px;z-index:1000;box-shadow:0 2px 5px rgba(0,0,0,0.2);';
-        
-        const messageElement = document.createElement('span');
-        messageElement.textContent = message;
-        
-        const closeButton = document.createElement('button');
-        closeButton.innerHTML = '&times;';
-        closeButton.style.cssText = 'margin-left:10px;background:none;border:none;color:white;font-size:16px;cursor:pointer;';
-        closeButton.addEventListener('click', () => {
-            container.style.display = 'none';
-        });
-        
-        container.appendChild(messageElement);
-        container.appendChild(closeButton);
-        document.body.appendChild(container);
-    } else {
-        // 如果存在，更新消息
-        const messageElement = errorContainer.querySelector('span') || errorContainer;
-        messageElement.textContent = message;
-        errorContainer.style.display = 'block';
-    }
-}
-
-/**
  * 显示欢迎消息
  */
 function showWelcomeMessage() {
@@ -521,82 +317,27 @@ function showUpdateMessage(oldVersion, newVersion) {
 }
 
 /**
- * 显示通知
- * @param {string} title - 通知标题
- * @param {string} message - 通知内容
- * @param {number} [duration=5000] - 显示持续时间(毫秒)
- */
-function showNotification(title, message, duration = 5000) {
-    // 创建通知元素
-    const notification = document.createElement('div');
-    notification.classList.add('notification');
-    notification.style.cssText = 'position:fixed;bottom:20px;right:20px;width:300px;background:white;border-radius:5px;box-shadow:0 2px 10px rgba(0,0,0,0.2);overflow:hidden;z-index:1000;transform:translateY(100%);transition:transform 0.3s;';
-    
-    // 添加通知内容
-    notification.innerHTML = `
-        <div style="padding:15px;border-bottom:1px solid #eee;">
-            <div style="display:flex;justify-content:space-between;align-items:center;">
-                <h3 style="margin:0;font-size:16px;">${title}</h3>
-                <button class="close-btn" style="background:none;border:none;font-size:16px;cursor:pointer;">&times;</button>
-            </div>
-            <p style="margin:10px 0 0;font-size:14px;">${message}</p>
-        </div>
-    `;
-    
-    // 添加到文档
-    document.body.appendChild(notification);
-    
-    // 显示通知
-    setTimeout(() => {
-        notification.style.transform = 'translateY(0)';
-    }, 100);
-    
-    // 添加关闭按钮事件
-    const closeBtn = notification.querySelector('.close-btn');
-    closeBtn.addEventListener('click', () => {
-        closeNotification();
-    });
-    
-    // 定时关闭
-    const timeoutId = setTimeout(() => {
-        closeNotification();
-    }, duration);
-    
-    // 关闭通知函数
-    function closeNotification() {
-        notification.style.transform = 'translateY(100%)';
-        clearTimeout(timeoutId);
-        
-        // 动画结束后移除DOM
-        setTimeout(() => {
-            if (document.body.contains(notification)) {
-                document.body.removeChild(notification);
-            }
-        }, 300);
-    }
-}
-
-/**
- * 处理页面可见性变化
- * 当页面从隐藏变为可见时，可能需要刷新某些数据
- */
-function handleVisibilityChange() {
-    if (!document.hidden && isInitialized) {
-        // 如果页面变为可见且已初始化，执行刷新操作
-        refreshPageContent();
-    }
-}
-
-/**
  * 刷新页面内容
  * 用于在标签页重新激活时更新内容
  */
 async function refreshPageContent() {
     try {
-        // 这里可以添加需要刷新的内容
-        // 例如刷新背景图片、更新时间等
+        // 刷新背景图片
+        await refreshBackgroundImage();
+        // 更新时钟
+        updateClock();
     } catch (error) {
         console.error('Failed to refresh page content:', error);
+    }
+}
+
+/**
+ * 处理页面可见性变化
+ */
+function handleVisibilityChange() {
+    if (!document.hidden && isInitialized) {
+        // 如果页面变为可见且已初始化，执行刷新操作
+        refreshPageContent();
     }
 }
 
@@ -606,9 +347,8 @@ document.addEventListener('visibilitychange', handleVisibilityChange);
 // 页面加载完成后初始化应用
 document.addEventListener('DOMContentLoaded', init);
 
-// 在 newtab.js 文件中添加
+// 加载超时保护
 document.addEventListener('DOMContentLoaded', () => {
-  // 加载超时保护
   setTimeout(function() {
     const loadingScreen = document.getElementById('loading-screen');
     if (loadingScreen && loadingScreen.style.display !== 'none') {
@@ -620,19 +360,28 @@ document.addEventListener('DOMContentLoaded', () => {
     }
   }, 10000);
 });
+
+// 在开始初始化前，设置一个绝对的超时保护
 document.addEventListener('DOMContentLoaded', () => {
-    // 首先初始化核心功能
-    initCoreFeatures();
-    
-    // 延迟加载非核心功能
+  // 保持现有的超时保护
+  setTimeout(function() {
+    const loadingScreen = document.getElementById('loading-screen');
+    if (loadingScreen && loadingScreen.style.display !== 'none') {
+      console.warn('Loading timeout - force hiding loading screen');
+      // 使用立即模式强制隐藏
+      hideLoadingIndicator(true);
+    }
+  }, 8000); // 缩短到8秒，确保不会等待太久
+  
+  // 添加备份保护机制，确保页面加载后显示内容
+  window.addEventListener('load', () => {
     setTimeout(() => {
-        initNonCriticalFeatures();
-    }, 500);
+      hideLoadingIndicator(true);
+    }, 2000); // 页面加载完成2秒后强制隐藏
+  });
 });
 
 // 导出一些可能在外部使用的函数
 export {
-    VERSION,
-    showNotification,
-    showErrorMessage
+    VERSION
 };
